Update code with PR review comments
